refactor: Use nested classes
