+++ lib/storage_tables/service.rb
@@ -5,14 +5,14 @@ module StorageTables
   class Service
     extend ActiveSupport::Autoload
 
-    autoload StorageTables::Services::Configurator
+    autoload StorageTables::Service::Configurator
 
     class << self
       # Configure an Active Storage service by name from a set of configurations,
       # typically loaded from a YAML file. The Active Storage engine uses this
       # to set the global Active Storage service when the app boots.
       def configure(service_name, configurations)
-        StorageTables::Services::Configurator.build(service_name, configurations)
+        StorageTables::Service::Configurator.build(service_name, configurations)
       end
 
       private
