+++ lib/storage_tables/service/configurator.rb
@@ -4,14 +4,14 @@ require "active_storage/service"
 require "active_storage/service/configurator"
 
 module StorageTables
-  module Services
+  class Service
     # Set the storage service to be used by Storage Tables.
     class Configurator < ActiveStorage::Service::Configurator
       private
 
       def resolve(class_name)
-        require "storage_tables/services/#{class_name.to_s.underscore}_service"
-        StorageTables::Services.const_get(:"#{class_name.camelize}Service")
+        require "storage_tables/service/#{class_name.to_s.underscore}_service"
+        StorageTables::Service.const_get(:"#{class_name.camelize}Service")
       rescue LoadError
         raise "Missing service adapter for #{class_name.inspect}"
       end
