+++ lib/storage_tables/service/disk_service.rb
@@ -3,7 +3,7 @@
 require "active_storage/service/disk_service"
 
 module StorageTables
-  module Services
+  class Service
     # Local disk storage service.
     class DiskService < ActiveStorage::Service::DiskService
       def path_for(checksum) # :nodoc:
