+++ lib/storage_tables/service/registry.rb
@@ -2,15 +2,15 @@
 
 require "active_storage/service"
 require "active_storage/service/registry"
-require "storage_tables/services/configurator"
+require "storage_tables/service/configurator"
 
 module StorageTables
-  module Services
+  class Service
     class Registry < ActiveStorage::Service::Registry # :nodoc:
       private
 
       def configurator
-        @configurator ||= StorageTables::Services::Configurator.new(configurations)
+        @configurator ||= StorageTables::Service::Configurator.new(configurations)
       end
     end
   end
