+++ test/service/shared_service_tests.rb
@@ -4,7 +4,7 @@ require "test_helper"
 require "active_support/core_ext/securerandom"
 
 module StorageTables
-  module Services
+  class Service
     module SharedServiceTests
       extend ActiveSupport::Concern
